code/deployment/app/requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
pydantic==2.9.2